# solution/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AdsPlatform API",
    "version": "0.1.0"
  },
  "paths": {
    "/s3/{file_path}": {
      "get": {
        "tags": [
          "S3"
        ],
        "summary": "Get File",
        "operationId": "get_file_s3__file_path__get",
        "parameters": [
          {
            "name": "file_path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "File Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "S3"
        ],
        "summary": "Upload File",
        "operationId": "upload_file_s3__file_path__put",
        "parameters": [
          {
            "name": "file_path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "File Path"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_file_s3__file_path__put"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "S3"
        ],
        "summary": "Delete File",
        "operationId": "delete_file_s3__file_path__delete",
        "parameters": [
          {
            "name": "file_path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "File Path"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/advertisers/{advertiser_id}/campaigns": {
      "post": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Create Campaign",
        "operationId": "create_campaign_advertisers__advertiser_id__campaigns_post",
        "parameters": [
          {
            "name": "advertiser_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Advertiser Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CampaignIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Campaign"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Get Campaigns",
        "operationId": "get_campaigns_advertisers__advertiser_id__campaigns_get",
        "parameters": [
          {
            "name": "advertiser_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Advertiser Id"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Size"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "default": 0,
              "title": "Page"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Campaign"
                  },
                  "title": "Response Get Campaigns Advertisers  Advertiser Id  Campaigns Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/stats/campaigns/{campaign_id}": {
      "get": {
        "tags": [
          "Statistics"
        ],
        "summary": "Campaign Stats",
        "operationId": "campaign_stats_stats_campaigns__campaign_id__get",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignStats"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/stats/advertisers/{advertiser_id}/campaigns": {
      "get": {
        "tags": [
          "Statistics"
        ],
        "summary": "Advertiser Stats",
        "operationId": "advertiser_stats_stats_advertisers__advertiser_id__campaigns_get",
        "parameters": [
          {
            "name": "advertiser_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Advertiser Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignStats"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/stats/campaigns/{campaign_id}/daily": {
      "get": {
        "tags": [
          "Statistics"
        ],
        "summary": "Campaign Daily Stats",
        "operationId": "campaign_daily_stats_stats_campaigns__campaign_id__daily_get",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CampaignDailyStats"
                  },
                  "title": "Response Campaign Daily Stats Stats Campaigns  Campaign Id  Daily Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/stats/advertisers/{advertiser_id}/campaigns/daily": {
      "get": {
        "tags": [
          "Statistics"
        ],
        "summary": "Advertiser Daily Stats",
        "operationId": "advertiser_daily_stats_stats_advertisers__advertiser_id__campaigns_daily_get",
        "parameters": [
          {
            "name": "advertiser_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Advertiser Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CampaignDailyStats"
                  },
                  "title": "Response Advertiser Daily Stats Stats Advertisers  Advertiser Id  Campaigns Daily Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_file_s3__file_path__put": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_file_s3__file_path__put"
      },
      "Campaign": {
        "properties": {
          "impressions_limit": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Impressions Limit"
          },
          "clicks_limit": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Clicks Limit"
          },
          "cost_per_impression": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Cost Per Impression"
          },
          "cost_per_click": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Cost Per Click"
          },
          "start_date": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Start Date"
          },
          "end_date": {
            "type": "integer",
            "minimum": 0.0,
            "title": "End Date"
          },
          "ad_title": {
            "type": "string",
            "title": "Ad Title"
          },
          "ad_text": {
            "type": "string",
            "title": "Ad Text"
          },
          "ad_image": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ad Image"
          },
          "targeting": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CampaignTargeting"
              },
              {
                "type": "null"
              }
            ]
          },
          "campaign_id": {
            "type": "string",
            "format": "uuid",
            "title": "Campaign Id"
          },
          "advertiser_id": {
            "type": "string",
            "format": "uuid",
            "title": "Advertiser Id"
          }
        },
        "type": "object",
        "required": [
          "impressions_limit",
          "clicks_limit",
          "cost_per_impression",
          "cost_per_click",
          "start_date",
          "end_date",
          "ad_title",
          "ad_text",
          "campaign_id",
          "advertiser_id"
        ],
        "title": "Campaign"
      },
      "CampaignDailyStats": {
        "properties": {
          "impressions_count": {
            "type": "integer",
            "title": "Impressions Count"
          },
          "clicks_count": {
            "type": "integer",
            "title": "Clicks Count"
          },
          "conversion": {
            "type": "number",
            "title": "Conversion"
          },
          "spent_impressions": {
            "type": "number",
            "title": "Spent Impressions"
          },
          "spent_clicks": {
            "type": "number",
            "title": "Spent Clicks"
          },
          "spent_total": {
            "type": "number",
            "title": "Spent Total"
          },
          "date": {
            "type": "integer",
            "title": "Date"
          }
        },
        "type": "object",
        "required": [
          "impressions_count",
          "clicks_count",
          "conversion",
          "spent_impressions",
          "spent_clicks",
          "spent_total",
          "date"
        ],
        "title": "CampaignDailyStats"
      },
...
      "CampaignIn": {
        "properties": {
          "impressions_limit": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Impressions Limit"
          },
          "clicks_limit": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Clicks Limit"
          },
          "cost_per_impression": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Cost Per Impression"
          },
          "cost_per_click": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Cost Per Click"
          },
          "start_date": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Start Date"
          },
          "end_date": {
            "type": "integer",
            "minimum": 0.0,
            "title": "End Date"
          },
          "ad_title": {
            "type": "string",
            "title": "Ad Title"
          },
          "ad_text": {
            "type": "string",
            "title": "Ad Text"
          },
          "ad_image": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ad Image"
          },
          "targeting": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CampaignTargeting"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "impressions_limit",
          "clicks_limit",
          "cost_per_impression",
          "cost_per_click",
          "start_date",
          "end_date",
          "ad_title",
          "ad_text"
        ],
        "title": "CampaignIn"
      },
      "CampaignStats": {
        "properties": {
          "impressions_count": {
            "type": "integer",
            "title": "Impressions Count"
          },
          "clicks_count": {
            "type": "integer",
            "title": "Clicks Count"
          },
          "conversion": {
            "type": "number",
            "title": "Conversion"
          },
          "spent_impressions": {
            "type": "number",
            "title": "Spent Impressions"
          },
          "spent_clicks": {
            "type": "number",
            "title": "Spent Clicks"
          },
          "spent_total": {
            "type": "number",
            "title": "Spent Total"
          }
        },
        "type": "object",
        "required": [
          "impressions_count",
          "clicks_count",
          "conversion",
          "spent_impressions",
          "spent_clicks",
          "spent_total"
        ],
        "title": "CampaignStats"
      },
      "CampaignTargeting": {
        "properties": {
          "gender": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TargetingGender"
              },
              {
                "type": "null"
              }
            ]
          },
          "age_from": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Age From"
          },
          "age_to": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Age To"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          }
        },
        "type": "object",
        "title": "CampaignTargeting"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TargetingGender": {
        "type": "string",
        "enum": [
          "MALE",
          "FEMALE",
          "ALL"
        ],
        "title": "TargetingGender"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}